finished keyword searcher fully
