+++ backend/routes/getFromMongo.ts
@@ -1,6 +1,7 @@
 // getFromMongo.ts
 import express, { Router, Request, Response } from 'express';
 import { PrismaClient } from '../generated/prisma';
+import { filterKeywords } from '../services/keywordService';
 
 const router: Router = express.Router();
 const prisma = new PrismaClient();
@@ -87,16 +88,66 @@ router.get('/user-audio-files', async (_req: Request, res: Response) => {
  * GET /mongo/audio-file-keywords
  * Returns all AudioFileKeyword entries + their parent audioFile
  */
-router.get('/audio-file-keywords', async (_req: Request, res: Response) => {
+// router.get('/audio-file-keywords', async (_req: Request, res: Response) => {
+//   try {
+//     const keywords = await prisma.audioFileKeyword.findMany({
+//       include: { audioFile: true }
+//     });
+//     res.json(keywords);
+//   } catch (err: any) {
+//     console.error(err);
+//     res.status(500).json({ error: err.message });
+//   }
+// });
+
+// --------------------------------------------------------------------
+// GET /generate-podcast?prompt=...&firebaseId=...
+// – Looks for an existing podcast whose keyword set overlaps with
+//   the keywords extracted from the prompt.
+// – If found, returns it (including the audio URL) and *does not*
+//   create anything new.
+// – If not found, responds with 404 so that the client can fall
+//   back to the POST route to create a brand-new episode.
+// --------------------------------------------------------------------
+router.get('/audio-file-by-keywords', async (req: Request, res: Response): Promise<void> => {
   try {
-    const keywords = await prisma.audioFileKeyword.findMany({
-      include: { audioFile: true }
+    const prompt     = String(req.query.prompt ?? '').trim();
+    const firebaseId = String(req.query.firebaseId ?? '').trim();
+
+    if (!prompt || !firebaseId) {
+      res.status(400).json({ error: 'Both prompt and firebaseId are required' });
+      return;
+    }
+
+    // 1️⃣  Extract keywords from the user prompt
+    const keywords = filterKeywords(prompt);   // e.g. ["React", "virtual DOM"]
+
+    // 2️⃣  Look for a matching summary that shares *any* of those keywords
+    //      and eagerly load its linked audio file (1-to-1 relation).
+    const existing = await prisma.podcastSummary.findFirst({
+      where: { keywords: { hasSome: keywords } },
+      include: { audio: true },                // pull the AudioFile row too
     });
-    res.json(keywords);
-  } catch (err: any) {
-    console.error(err);
-    res.status(500).json({ error: err.message });
+
+    // 3️⃣  If we found one, return it; otherwise signal “not found”
+    if (existing && existing.audio) {
+      res.json({
+        id:        existing.id,
+        title:     existing.title,
+        content:   existing.content,
+        summary:   existing.summary,
+        keywords:  existing.keywords,
+        audioUrl:  existing.audio.url,   // public / signed S3 URL
+        s3Key:     existing.audio.s3Key,
+      });
+    } else {
+      res.status(404).json({ message: 'No cached podcast found' });
+    } 
+  } catch (error) {
+    console.error('GET /generate-podcast error:', error);
+    res.status(500).json({ error: 'Failed to look up podcast' });
   }
 });
+
 
 export default router;
